[#10] - UI/UX improvments, unrealted to 10 but its in here
