+++ src/components/base/LinkHoverPreview.vue
@@ -73,5 +73,3 @@ onUnmounted(() => {
   }
 });
 </script>
-
-
