+++ src/components/base/SettingsSection.vue
@@ -1,0 +1,21 @@
+<template>
+  <div class="settings-section">
+    <div class="section-header">
+      <h2>{{ title }}</h2>
+      <div v-if="$slots.actions" class="section-actions">
+        <slot name="actions" />
+      </div>
+    </div>
+    <div class="section-content">
+      <slot />
+    </div>
+  </div>
+</template>
+
+<script setup lang="ts">
+interface Props {
+  title: string;
+}
+
+defineProps<Props>();
+</script>
