+++ src/components/views/CustomView.vue
@@ -89,10 +89,10 @@
 
 <script setup lang="ts">
 import { ref, computed, onMounted, watch } from 'vue';
-import { useItems } from '../composables/useItems';
-import { useCustomViews } from '../composables/useCustomViews';
-import type { CustomView } from '../types';
-import { formatDate, truncate, stripHtml, parseGroups } from '../utils/formatting';
+import { useItems } from '../../composables/useItems';
+import { useCustomViews } from '../../composables/useCustomViews';
+import type { CustomView } from '../../types';
+import { formatDate, truncate, stripHtml, parseGroups } from '../../utils/formatting';
 
 const props = defineProps<{
   viewId: number;
