+++ src/components/views/TodayView.vue
@@ -131,9 +131,9 @@
 
 <script setup lang="ts">
 import { ref, computed, onMounted, watch } from 'vue';
-import { useItems } from '../composables/useItems';
-import { useSources } from '../composables/useSources';
-import { formatDate, truncate, stripHtml, parseGroups } from '../utils/formatting';
+import { useItems } from '../../composables/useItems';
+import { useSources } from '../../composables/useSources';
+import { formatDate, truncate, stripHtml, parseGroups } from '../../utils/formatting';
 
 const emit = defineEmits<{
   (e: 'select-item', id: number): void;
