+++ src/styles/components/_settings-section.scss
@@ -1,0 +1,76 @@
+// SettingsSection Component Styles
+@use '../variables' as *;
+@use '../mixins' as *;
+
+.settings-section {
+  margin-bottom: $spacing-3xl;
+  width: 100%;
+  
+  .section-header {
+    display: flex;
+    justify-content: space-between;
+    align-items: center;
+    gap: $spacing-md;
+    margin-bottom: $spacing-xl;
+    width: 100%;
+    box-sizing: border-box;
+    padding-bottom: $spacing-md;
+    border-bottom: 1px solid var(--color-border);
+    
+    h2 {
+      margin: 0;
+      font-size: $font-xl;
+      font-weight: $font-semibold;
+      color: var(--color-text-primary);
+      white-space: nowrap;
+    }
+    
+    .section-actions {
+      display: flex;
+      gap: $spacing-sm;
+      flex-wrap: nowrap;
+      justify-content: flex-end;
+      
+      .action-button {
+        white-space: nowrap;
+        padding: $spacing-sm $spacing-lg;
+        font-size: $font-base;
+        border-radius: $radius-md;
+        font-weight: $font-medium;
+        transition: all $transition-base;
+        
+        &.primary {
+          background: var(--color-primary);
+          color: var(--color-primary-text);
+          border: 1px solid var(--color-primary);
+          
+          &:hover:not(:disabled) {
+            opacity: 0.9;
+            transform: translateY(-1px);
+            box-shadow: 0 2px 8px rgba(var(--color-primary-rgb), 0.3);
+          }
+        }
+        
+        &:not(.primary) {
+          background: var(--color-bg-secondary);
+          color: var(--color-text-primary);
+          border: 1px solid var(--color-border);
+          
+          &:hover:not(:disabled) {
+            background: var(--color-bg-hover);
+            border-color: var(--color-primary);
+          }
+        }
+        
+        &:disabled {
+          opacity: 0.5;
+          cursor: not-allowed;
+        }
+      }
+    }
+  }
+  
+  .section-content {
+    width: 100%;
+  }
+}
