+++ src/styles/components/_tabbed-view.scss
@@ -1,0 +1,92 @@
+// TabbedView Component Styles
+@use '../variables' as *;
+@use '../mixins' as *;
+
+.tabbed-view {
+  width: 100%;
+}
+
+// Tab Navigation - Page-style tabs (like browser tabs)
+.tabs-nav {
+  display: flex;
+  gap: $spacing-xs;
+  margin-bottom: $spacing-xl;
+  border-bottom: 2px solid var(--color-border);
+  position: relative;
+  z-index: 10;
+  background: var(--color-bg-primary);
+  padding: 0;
+  margin-top: $spacing-lg;
+  padding-bottom: 0;
+}
+
+.tab-button {
+  display: flex;
+  align-items: center;
+  gap: $spacing-sm;
+  padding: $spacing-md $spacing-xl;
+  border: 1px solid var(--color-border);
+  border-bottom: none;
+  background: var(--color-bg-secondary);
+  color: var(--color-text-secondary);
+  font-size: $font-base;
+  font-weight: $font-medium;
+  cursor: pointer;
+  border-radius: $radius-md $radius-md 0 0;
+  transition: all $transition-base;
+  position: relative;
+  white-space: nowrap;
+  margin-bottom: -2px;
+  
+  .tab-icon {
+    font-size: $font-lg;
+    line-height: 1;
+  }
+  
+  &:hover:not(.active) {
+    background: var(--color-bg-hover);
+    color: var(--color-text-primary);
+    border-color: var(--color-border);
+  }
+  
+  &.active {
+    color: var(--color-primary);
+    background: var(--color-bg-primary);
+    font-weight: $font-semibold;
+    border-color: var(--color-border);
+    border-bottom-color: var(--color-bg-primary);
+    z-index: 1;
+    box-shadow: 0 -2px 8px rgba(0, 0, 0, 0.05);
+    
+    &::after {
+      content: '';
+      position: absolute;
+      bottom: -2px;
+      left: 0;
+      right: 0;
+      height: 2px;
+      background: var(--color-bg-primary);
+    }
+  }
+}
+
+.tab-content {
+  min-height: 400px;
+  padding-top: $spacing-lg;
+}
+
+.tab-pane {
+  animation: fadeIn 0.2s ease-in-out;
+  width: 100%;
+}
+
+@keyframes fadeIn {
+  from {
+    opacity: 0;
+    transform: translateY(4px);
+  }
+  to {
+    opacity: 1;
+    transform: translateY(0);
+  }
+}
